+++ src/main/java/com/jbucloud/festival/quizgame/repository/GameSessionRepository.java
@@ -1,6 +1,6 @@
-package com.jbucloud.festival.repository;
+package com.jbucloud.festival.quizgame.repository;
 
-import com.jbucloud.festival.domain.GameSession;
+import com.jbucloud.festival.quizgame.domain.GameSession;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
